requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.4
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
pyee==7.0.4
PyJWT==2.0.0
pyppeteer==0.2.2
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
sqlparse==0.4.1
tqdm==4.55.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
webdriver-manager==3.2.2
websockets==8.1
whitenoise==5.2.0
yarl==1.6.3
